@@ -15,15 +15,32 @@
     "css:watch": "npm run build:css -- --w",
     "dev": "pm2-dev ./other/pm2.config.js",
     "postinstall": "remix setup",
+    "prepare": "husky install",
     "setup": "npm install && docker compose up -d",
-    "start": "cross-env NODE_ENV=production node --require ./node_modules/dotenv/config ./index.js"
+    "start": "cross-env NODE_ENV=production node --require ./node_modules/dotenv/config ./index.js",
+    "lint": "eslint --cache --cache-location ./node_modules/.cache/.eslintcache --ext js,jsx,ts,tsx .",
+    "format": "prettier --write \"**/*.+(js|jsx|json|yml|yaml|css|less|scss|ts|tsx|md|gql|graphql|mdx|vue)\"",
+    "typecheck": "tsc -b",
+    "test": "jest --passWithNoTests",
+    "validate": "./other/validate"
   },
+  "eslintIgnore": [
+    "node_modules",
+    "coverage",
+    "server-build",
+    "build",
+    "public/build",
+    "*.ignored/",
+    "*.ignored.*"
+  ],
   "dependencies": {
     "@octokit/plugin-throttling": "^3.5.2",
     "@octokit/rest": "^18.12.0",
     "@prisma/client": "^3.6.0",
-    "@remix-run/react": "^1.0.5",
+    "@remix-run/express": "^1.0.6",
+    "@remix-run/react": "^1.0.6",
     "@remix-run/serve": "^1.0.5",
+    "@sentry/node": "^6.16.1",
     "@tailwindcss/typography": "^0.4.1",
     "cloudinary-build-url": "^0.2.1",
     "clsx": "^1.1.1",
@@ -34,6 +51,8 @@
     "error-stack-parser": "^2.0.6",
     "esbuild": "^0.13.14",
     "express": "^4.17.1",
+    "fs-extra": "^10.0.0",
+    "glob": "^7.2.0",
     "lodash.isequal": "^4.5.0",
     "mdx-bundler": "^8.0.0",
     "morgan": "^1.10.0",
@@ -55,6 +74,11 @@
   "devDependencies": {
     "@cld-apis/types": "^0.1.3",
     "@remix-run/dev": "^1.0.5",
+    "@testing-library/cypress": "^8.0.2",
+    "@testing-library/jest-dom": "^5.15.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/react-hooks": "^7.0.2",
+    "@testing-library/user-event": "^13.5.0",
     "@types/compression": "^1.7.2",
     "@types/express": "^4.17.13",
     "@types/lodash.isequal": "^4.5.5",
@@ -64,14 +88,24 @@
     "@types/react-dom": "^17.0.9",
     "@types/redis": "^2.8.32",
     "autoprefixer": "^10.4.0",
+    "concurrently": "^6.4.0",
     "dotenv-cli": "^4.1.1",
+    "esbuild-jest": "^0.5.0",
     "esbuild-register": "^3.2.0",
+    "eslint": "^7.32.0",
+    "eslint-config-kentcdodds": "^19.2.0",
+    "husky": "^7.0.4",
+    "jest": "^27.4.5",
+    "jest-watch-typeahead": "^1.0.0",
+    "kcd-scripts": "^11.2.2",
     "postcss": "^8.3.11",
     "postcss-cli": "^9.0.2",
     "postcss-import": "^14.0.2",
+    "prettier": "^2.5.0",
+    "prettier-plugin-tailwind": "^2.2.12",
     "rimraf": "^3.0.2",
     "tailwindcss": "^2.2.19",
-    "typescript": "^4.1.2"
+    "typescript": "4.4.4"
   },
   "engines": {
     "node": "16",
