@@ -6,13 +6,19 @@ type Props = LinkProps & {
   external?: boolean
 }
 
-export default function Link({external, className, to, ...rest}: Props) {
+export default function Link({
+  external,
+  className,
+  to,
+  children,
+  ...rest
+}: Props) {
   const classes = clsx(
-    'text-gray-600 hover:text-gray-500 dark:text-gray-400 dark:hover:text-gray-500 transition',
+    'dark:hover:text-gray-500 dark:text-gray-400 hover:text-gray-500 text-gray-600 transition',
     className,
   )
 
-  const href = typeof to === 'string' ? to : to?.pathname
+  const href = typeof to === 'string' ? to : to.pathname
 
   if (external) {
     return (
@@ -22,9 +28,15 @@ export default function Link({external, className, to, ...rest}: Props) {
         target="_blank"
         rel="noopener noreferrer"
         className={classes}
-      />
+      >
+        {children}
+      </a>
     )
   }
 
-  return <RemixLink {...rest} to={to} className={classes} />
+  return (
+    <RemixLink {...rest} to={to} className={classes}>
+      {children}
+    </RemixLink>
+  )
 }
