@@ -1,4 +1,4 @@
 export default [
-  {name: 'Home', to: '/'},
-  {name: 'Blog', to: '/blog'},
+  {name: '🏠 Home', to: '/'},
+  {name: '📰 Blog', to: '/blog'},
 ]
