@@ -11,11 +11,11 @@ export default function MenuItem({
     to === location.pathname || location.pathname.startsWith(`${to}/`)
 
   return (
-    <li className="border-b border-gray-200 dark:border-gray-600 text-gray-900 dark:text-gray-100 text-sm font-semibold w-full">
+    <li className="w-full dark:text-gray-100 text-gray-900 text-sm font-semibold border-b border-gray-200 dark:border-gray-600">
       <Link
         prefetch="intent"
         className={clsx(
-          'flex w-auto p-6 hover:bg-gray-200 dark:hover:bg-gray-800',
+          'dark:hover:bg-gray-800 flex p-6 w-auto hover:bg-gray-200',
           {
             'font-semibold text-gray-800 dark:text-gray-200': isSelected,
             'font-normal text-gray-600 dark:text-gray-400': !isSelected,
