@@ -4,7 +4,7 @@ export default function SkipNav() {
   return (
     <a
       href="#main"
-      className="text-primary bg-primary fixed z-10 px-4 py-3 transition-transform duration-200 transform -translate-y-12 left-2 focus:top-4 focus:translate-y-3 -top-8"
+      className="text-primary bg-primary fixed z-10 -top-8 left-2 focus:top-4 px-4 py-3 transform -translate-y-12 focus:translate-y-3 transition-transform duration-200"
     >
       Skip to content
     </a>
