@@ -2,6 +2,7 @@ import {PrismaClient} from '@prisma/client'
 import {getRequiredServerEnvVar} from './misc'
 
 declare global {
+  // eslint-disable-next-line no-var,vars-on-top
   var prisma: ReturnType<typeof getClient> | undefined
 }
 
@@ -10,6 +11,7 @@ declare global {
  *
  */
 // @ts-expect-error ts(2339)
+// eslint-disable-next-line no-extend-native
 BigInt.prototype.toJSON = function toJSON() {
   return Number(this)
 }
@@ -36,7 +38,7 @@ function getClient(connectionUrl: URL): PrismaClient {
       },
     },
   })
-  client.$on('query', (e: {duration: number; query: any}) => {
+  client.$on('query', (e: {duration: number; query: unknown}) => {
     if (e.duration < logThreshold) return
 
     const dur = `${e.duration}ms`
